// munch/Week4/Week4_Mission1/src/types/movieDetail.ts
export interface CrewMember {
	id: number;
	name: string;
	job: string;
	department: string;
	profile_path: string | null;
}

export interface MovieDetail {
	id: number;
	title: string;
	original_title: string;
	overview: string;
	poster_path: string;
	backdrop_path: string;
	release_date: string;
	runtime: number;
	vote_average: number;
	vote_count: number;
	tagline: string;
	status: string;
	genres: { id: number; name: string }[];
	production_companies: {
		id: number;
		name: string;
		logo_path: string | null;
	}[];
	popularity: number;
	adult: boolean;
}

export interface Credits {
	cast: {
		id: number;
		name: string;
		character: string;
		profile_path: string | null;
		order: number;
	}[];
	crew: CrewMember[];
}